unit testing bug fixes - leave event

+++ wattpad/pluginsexecution/eventsexecution/eventexec.py
@@ -17,7 +17,7 @@ class Eventexec:
         self.channelRepo= ChannelRepo()
         self.customMsgRepo= CustomMsgrepo()
 
-    async def guild_join_event(self, guildId: str) -> None:
+    async def guild_join_event(self, guildId: str) -> bool:
         try:
             self.logger.info("%s.guild_join_event method invoked for server: %s", self.file_prefix, guildId)
 
@@ -25,15 +25,17 @@ class Eventexec:
             server=Server(GuildId=guildId, IsActive=1)  
             result= await self.serverRepo.insert_server_data(server)  
 
-            if not result:
-                self.logger.fatal("Exception occured in %s.guild_join_event while inserting server data for server: %s", self.file_prefix, guildId)
+            if result:
+                return True
 
+
+            self.logger.fatal("Exception occured in %s.guild_join_event while inserting server data for server: %s", self.file_prefix, guildId)                
         
         except Exception as e:
             self.logger.fatal("Exception occured in %s.guild_join_event method invoked for server: %s", self.file_prefix, guildId,exc_info=1)
             raise e
         
-    async def guild_leave_event(self, guildId: str) -> None:
+    async def guild_leave_event(self, guildId: str) -> bool:
         try:
             self.logger.info("%s.guild_join_event method invoked for server: %s", self.file_prefix, guildId)
 
@@ -45,34 +47,40 @@ class Eventexec:
 
             serverid= await self.serverRepo.get_serverid_from_server(guildId)
 
-            stories= await self.storyRepo.get_stories_from_server_id(serverid, 1)
-            authors= await self.authorRepo.get_authors_from_server_id(serverid, 1)
+            if serverid:
+                stories= await self.storyRepo.get_stories_from_server_id(serverid, 1)
+                authors= await self.authorRepo.get_authors_from_server_id(serverid, 1)
 
-            story_inactivate_result= await self.storyRepo.inactivate_all_stories_by_server_id(serverid)
-            author_inactivate_result= await self.authorRepo.inactivate_all_authors_by_server_id(serverid)
+                story_inactivate_result= await self.storyRepo.inactivate_all_stories_by_server_id(serverid)
+                author_inactivate_result= await self.authorRepo.inactivate_all_authors_by_server_id(serverid)
 
-            if not story_inactivate_result:
-                self.logger.error("Error occured while inactivating stories for server: %s", serverid)
+                if not story_inactivate_result:
+                    self.logger.error("Error occured while inactivating stories for server: %s", serverid)
+                    return False
 
-            if not author_inactivate_result:
-                self.logger.error("Error occured while inactivating authors for server: %s", serverid)
+                if not author_inactivate_result:
+                    self.logger.error("Error occured while inactivating authors for server: %s", serverid)
+                    return False
 
-            for story in stories:
-                custom_msg_id= await self.customMsgRepo.get_custom_msg_id_from_story_id(story.StoryId, 1)
-                story_custom_msg_inactivate_result= await self.customMsgRepo.delete_custom_msg_by_id(custom_msg_id)
+                if stories:
+                    for story in stories:
+                        custom_msg_id= await self.customMsgRepo.get_custom_msg_id_from_story_id(story.StoryId, 1)
+                        story_custom_msg_inactivate_result= await self.customMsgRepo.delete_custom_msg_by_id(custom_msg_id)
 
-                if not story_custom_msg_inactivate_result:
-                    self.logger.error("Error occured while inactivating story custom msg: %s", custom_msg_id)
+                        if not story_custom_msg_inactivate_result:
+                            self.logger.error("Error occured while inactivating story custom msg: %s", custom_msg_id)
+                            return False
 
+                if authors:
+                    for author in authors:
+                        custom_msg_id= await self.customMsgRepo.get_custom_msg_id_from_author_id(author.AuthorId, 1)
+                        author_custom_msg_inactivate_result= await self.customMsgRepo.delete_custom_msg_by_id(custom_msg_id)
 
-            for author in authors:
-                custom_msg_id= await self.customMsgRepo.get_custom_msg_id_from_author_id(author.AuthorId, 1)
-                author_custom_msg_inactivate_result= await self.customMsgRepo.delete_custom_msg_by_id(custom_msg_id)
+                        if not author_custom_msg_inactivate_result:
+                            self.logger.error("Error occured while inactivating author custom msg: %s", custom_msg_id)
+                            return False
 
-                if not author_custom_msg_inactivate_result:
-                    self.logger.error("Error occured while inactivating author custom msg: %s", custom_msg_id)
-
-            
+            return True
         
         except Exception as e:
             self.logger.fatal("Exception occured in %s.guild_join_event method invoked for server: %s", self.file_prefix, guildId,exc_info=1)
